Move HintArea back to Hints.py
